[#179] Additional integration tests common logic
- stateless approach for template strings
